+++ src/fastapi_cloud_cli/commands/deploy/__init__.py
@@ -1,0 +1,3 @@
+from .command import deploy
+
+__all__ = ["deploy"]
